GpsLocationProvider: Reset GPS interface after cleanup
Since the init routine checks for the existence of the interface
struct (and skips execution if it exists), it was effectively only
being executed once. If the cleanup function in the GPS HAL destroys
data or handles that are established by init (which happens in the
libloc_api code), the end result is that GPS gets broken after being
disabled once, and can only be restored by a reboot.

This patch makes the HAL's init re-execute whenever GPS is (re)enabled.

Change-Id: I8eedbcd140484a132c5a90a06b7f95fbb0cab410